requirements.txt
docling
pandas
streamlit
watchdog
rapidocr_onnxruntime
ollama
opencv-python
numpy
torch
httpx
pillow